reorganize handlers hierarchically
