--- picopt/handlers/archive/__init__.py
@@ -1,1 +1,0 @@
-"""Archive Handlers."""
