+++ picopt/handlers/container/archive/seven_zip.py
@@ -10,7 +10,7 @@ from py7zr.py7zr import FileInfo as SevenZipInfo
 from termcolor import cprint
 
 from picopt.formats import FileFormat
-from picopt.handlers.archive.archive import PackingArchiveHandler
+from picopt.handlers.container.archive import PackingArchiveHandler
 from picopt.handlers.handler import INTERNAL
