+++ picopt/handlers/container/archive/tar.py
@@ -10,7 +10,7 @@ import filetype
 from termcolor import cprint
 
 from picopt.formats import FileFormat
-from picopt.handlers.archive.archive import PackingArchiveHandler
+from picopt.handlers.container.archive import PackingArchiveHandler
 from picopt.handlers.handler import INTERNAL
